+++ roocs_utils/xarray_utils/xarray_utils.py
@@ -12,11 +12,12 @@ from roocs_utils.project_utils import dset_to_filepaths
 known_coord_types = ["time", "level", "latitude", "longitude"]
 
 
-def open_xr_dataset(dset):
+def open_xr_dataset(dset, **kwargs):
     """
     Opens an xarray dataset from a dataset input.
 
     :param dset: (Str or Path) ds_id, directory path or file path ending in *.nc.
+    :param kwargs: Any further keyword arguments to include when opening the dataset. use_cftime=True and decode_timedelta=False are used by default, along with combine="by_coords" for open_mfdataset only.
     Any list will be interpreted as list of files
     """
 
@@ -27,10 +28,24 @@ def open_xr_dataset(dset):
 
     # if a list we want a multi-file dataset
     if len(dset) > 1:
-        return xr.open_mfdataset(dset, use_cftime=True, combine="by_coords")
+        return xr.open_mfdataset(
+            dset, use_cftime=True, combine="by_coords", decode_timedelta=False, **kwargs
+        )
     # if there is only one file we only need to call open_dataset
     else:
-        return xr.open_dataset(dset[0], use_cftime=True)
+        try:
+            return xr.open_dataset(
+                dset[0], use_cftime=True, decode_timedelta=False, **kwargs
+            )
+        # catch when kwargs only exist for open_mfdataset
+        except TypeError:
+            return xr.open_mfdataset(
+                dset,
+                use_cftime=True,
+                combine="by_coords",
+                decode_timedelta=False,
+                **kwargs,
+            )
 
 
 # from dachar